+++ cpm-svm/cpm-svm/svm.h
@@ -14,11 +14,11 @@ public:
     //enum {TRAIN, TEST};
     //class Exception {};
     //SVM();
-    void Train(const double* data, const int* resp, int n_samples, int n_vars,
+    void Train(const double* data, const long* resp, int n_samples, int n_vars,
                int row_step, int col_step,
                double lambda, double epsilon_abs, double epsilon_tol, int tMax);
     //Real Predict(const std::list<Pair>& sample) const;
-    void Predict(const double* data, int* pred,
+    void Predict(const double* data, long* pred,
                  int n_samples, int n_vars,
                  int row_step, int col_step) const;
     //Real CalcError(const Data& data, int type) const;
